Fix: Resolve OpenTelemetry compilation errors
- Remove ASP.NET Core instrumentation from metrics builder (traces only)
- Add missing Runtime and Process instrumentation packages
- Cast decimal to double for histogram recording
- Build now succeeds with no errors

+++ backend/Program.cs
@@ -427,7 +427,7 @@ app.MapPost("/api/jobs", async (HttpRequest req, SqliteConnection conn) =>
     await ins.ExecuteNonQueryAsync();
     
     OpenTelemetryConfig.JobCreated.Add(1);
-    if (quoted.HasValue) OpenTelemetryConfig.JobQuoteAmount.Record(quoted.Value);
+    if (quoted.HasValue) OpenTelemetryConfig.JobQuoteAmount.Record((double)quoted.Value);
     
     return Results.Ok(new { id, title, status, quotedPrice = quoted });
 });
